Refactor AbstractTestFrameworkDetector hierachy
