+++ subprojects/testing-jvm/src/main/java/org/gradle/api/internal/tasks/testing/testng/TestNGDetector.java
@@ -17,15 +17,8 @@ package org.gradle.api.internal.tasks.testing.testng;
 
 import org.gradle.api.internal.tasks.testing.detection.AbstractTestFrameworkDetector;
 import org.gradle.api.internal.tasks.testing.detection.ClassFileExtractionManager;
-import org.gradle.api.internal.tasks.testing.detection.TestClassVisitor;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.io.File;
 
 class TestNGDetector extends AbstractTestFrameworkDetector<TestNGTestClassDetector> {
-    private static final Logger LOGGER = LoggerFactory.getLogger(TestNGDetector.class);
-
     TestNGDetector(ClassFileExtractionManager classFileExtractionManager) {
         super(classFileExtractionManager);
     }
@@ -35,32 +28,8 @@ class TestNGDetector extends AbstractTestFrameworkDetector<TestNGTestClassDetect
         return new TestNGTestClassDetector(this);
     }
 
-    /**
-     * Uses a TestClassVisitor to detect whether the class in the testClassFile is a test class. <p/> If the class is not a test, this function will go up the inheritance tree to check if a parent
-     * class is a test class. First the package of the parent class is checked, if it is a java.lang or groovy.lang the class can't be a test class, otherwise the parent class is scanned. <p/> When a
-     * parent class is a test class all the extending classes are marked as test classes.
-     */
     @Override
-    protected boolean processTestClass(final File testClassFile, boolean superClass) {
-        final TestClassVisitor classVisitor = classVisitor(testClassFile);
-
-        boolean isTest = classVisitor.isTest();
-
-        if (!isTest) {
-            final String superClassName = classVisitor.getSuperClassName();
-
-            final File superClassFile = getSuperTestClassFile(superClassName);
-
-            if (superClassFile != null) {
-                isTest = processSuperClass(superClassFile);
-            } else {
-                LOGGER.debug("test-class-scan : failed to scan parent class {}, could not find the class file",
-                        superClassName);
-            }
-        }
-
-        publishTestClass(isTest, classVisitor, superClass);
-
-        return isTest;
+    protected boolean isKnownTestCaseClassName(String testCaseClassName) {
+        return false;
     }
 }
